+++ src/utils/hooks/assets.ts
@@ -1,9 +1,19 @@
 import { useMutation, useQuery } from "@tanstack/react-query";
-import { getLocalAsset, LocalAsset, saveLocalAsset } from "~utils/data/assets";
+import { getAssetUploadStatus, getLocalAsset, LocalAsset, saveLocalAsset } from "~utils/data/assets";
 import { useTeamIncidentsByEvent } from "./incident";
 import { HookQueryOptions } from "./robotevents";
 import { fetchAssetBlob, getAssetOriginalURL, getAssetPreviewURL } from "~utils/data/share";
 import { useMemo } from "react";
+
+export function useAssetUploadStatus(id?: string) {
+  return useQuery({
+    queryKey: ["assets", id, "uploadStatus"],
+    queryFn: async () => {
+      if (!id) return null;
+      return await getAssetUploadStatus(id);
+    },
+  });
+}
 
 export function useLocalAsset(id?: string | null) {
   return useQuery({
@@ -89,7 +99,7 @@ export function useAssetOriginalURL(
   options?: HookQueryOptions<string | null>
 ) {
   return useQuery({
-    queryKey: ["assets", id, "original"],
+    queryKey: ["assets", id, sku, "original"],
     queryFn: async () => {
       const local = await getLocalAsset(id);
       if (local) {